move task-domain into common:task-poc
