+++ common/task-poc/src/main/kotlin/cz/kotox/common/task/poc/domain/usecase/GetAllTasksUseCase.kt
@@ -1,7 +1,7 @@
-package cz.kotox.task.domain.api.usecase
+package cz.kotox.common.task.poc.domain.usecase
 
 import cz.kotox.common.task.poc.data.api.respository.TaskRepository
-import cz.kotox.task.domain.api.model.Task
+import cz.kotox.common.task.poc.domain.model.Task
 import cz.kotox.task.domain.api.factory.toTask
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
